+++ bolero/environment/__init__.py
@@ -1,4 +1,5 @@
-from .environment import Environment, ContextualEnvironment, SetContext
+from .environment import Environment, ContextualEnvironment
+from .set_context import SetContext
 from .objective_functions import ObjectiveFunction
 from .contextual_objective_functions import ContextualObjectiveFunction