+++ bolero/environment/test/test_contextual_objective_functions.py
@@ -1,8 +1,7 @@
 import numpy as np
 from bolero.environment.contextual_objective_functions import (
     CONTEXTUAL_FUNCTIONS, ContextualObjectiveFunction)
-from bolero.utils.testing import assert_raise_message
-from nose.tools import assert_less, assert_almost_equal
+from nose.tools import assert_less, assert_almost_equal, assert_raises_regexp
 
 
 def test_optimum():
@@ -21,8 +20,8 @@ def test_optimum():
 
 def test_input_validation():
     env = ContextualObjectiveFunction("Unknown", 2)
-    assert_raise_message(ValueError, "Unknown function", env.init)
+    assert_raises_regexp(ValueError, "Unknown function", env.init)
     env = ContextualObjectiveFunction("LinearContextualSphere", 0, 1)
-    assert_raise_message(ValueError, "Number of parameters", env.init)
+    assert_raises_regexp(ValueError, "Number of parameters", env.init)
     env = ContextualObjectiveFunction("LinearContextualSphere", 1, 0)
-    assert_raise_message(ValueError, "Number of context dimensions", env.init)
+    assert_raises_regexp(ValueError, "Number of context dimensions", env.init)