+++ bolero/utils/test/test_log.py
@@ -1,10 +1,10 @@
 from subprocess import call
 from bolero.utils.log import HideExtern
-from bolero.utils.testing import assert_raise_message
+from nose.tools import assert_raises_regexp
 
 
 def test_hide_extern():
-    assert_raise_message(ValueError, "Stream 'std' not in", HideExtern, "std")
+    assert_raises_regexp(ValueError, "Stream 'std' not in", HideExtern, "std")
     with HideExtern("stdout"):
         call(["echo", "will never be seen"])
     with HideExtern("stderr"):