+++ bolero/utils/testing.py
@@ -1,20 +1,9 @@
-from nose.tools import assert_in, assert_true
+from nose.tools import assert_true
 import bolero
 import pkgutil
 import inspect
 import os
 import pickle
-
-
-# Taken from scikit-learn
-def assert_raise_message(exception, message, function, *args, **kwargs):
-    """Helper function to test error messages in exceptions"""
-    try:
-        function(*args, **kwargs)
-        raise AssertionError("Should have raised %r" % exception(message))
-    except exception as e:
-        error_message = str(e)
-        assert_in(message, error_message)
 
 
 def assert_pickle(name, obj):